projects/snowflake-etl-pipeline/requirements.txt
snowflake-connector-python==3.6.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
sqlalchemy==2.0.25
